moving time helper functions to src.data_scraper.time_helpers.py
