+++ test/data_scraper/test_data_scraper.py
@@ -1,8 +1,9 @@
 from unittest import TestCase
-import pandas as pd
 import numpy as np
-import os
 import sys
+
+from src.data_scraper.time_helpers import get_current_timestamp
+
 sys.path[0] = sys.path[0].split('trading_bot')[0] + 'trading_bot'
 from src.data_scraper.data_scraper import DataScraper
 from src.utils import *
